+++ src/pycliarr/__init__.py
@@ -1,1 +1,2 @@
-__version__ = "1.0.19"
+__version__ = "1.0.20"
+from pycliarr.api import exceptions, radarr, sonarr